Move GUI into pytentiostat.gui subpackage
Relocate the GUI code and designer files from src/GUI/ into
src/pytentiostat/gui/ so the GUI is a proper importable subpackage, and
move the image assets to src/pytentiostat/pics/ (the code references them
via ../pics). Add a gui package with __init__, a refactored main() entry
point, and an app_setup.create_app helper that applies a consistent,
readable stylesheet across all windows. Update the .flake8 per-file-ignore
path to match the new location.

This is rebuilt on top of a synced upstream/main so it no longer reverts
the simulator feature or src/pytentiostat/main.py.
