+++ .flake8
@@ -12,4 +12,4 @@ max-line-length = 79
 # https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#labels-why-pycodestyle-warnings
 extend-ignore = E203
 per-file-ignores =
-    src/GUI/*:E501
+    src/pytentiostat/gui/*:E501
