+++ news/gui-restructure.rst
@@ -1,0 +1,27 @@
+**Added:**
+
+* ``pytentiostat.gui`` subpackage exposing the GUI as an importable Python
+  package, including an ``app_setup.create_app`` helper that applies a
+  consistent, readable stylesheet across all windows.
+
+**Changed:**
+
+* Moved the GUI code and designer files from ``src/GUI/`` into
+  ``src/pytentiostat/gui/`` (image assets moved to ``src/pytentiostat/pics/``)
+  and updated the internal imports accordingly.
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
