+++ src/pytentiostat/gui/CV_GUI.py
@@ -1,21 +1,22 @@
 # # Standard library
-from Adv_params_GUI import Ui_Adv_Params
+from PySide6 import QtCore, QtGui, QtWidgets
+
+from pytentiostat.gui.Adv_params_GUI import Ui_Adv_Params
 
 # # Local library
 # GUI_function
-from GUI_load_config import Ui_Load
-from PySide6 import QtCore, QtGui, QtWidgets
+from pytentiostat.gui.GUI_load_config import Ui_Load
 
 
-class Ui_CA(object):
+class Ui_CV(object):
+    """Initializes the 'Load config file' window.
+
+    Returns
+    ------
+    string : the loaded filename
+    """
 
     def load_folder_name(self):
-        """Initializes the 'Load config file' window.
-
-        Returns
-        ------
-        string : the loaded filename
-        """
         self.window = QtWidgets.QWidget()
         self.Load = Ui_Load()
         return self.Load.setupUi_save(self.window)
@@ -34,22 +35,42 @@ class Ui_CA(object):
         self.window.show()
         return self.AP, self.window
 
-    def setupUi(self, CA):
+    def setupUi(self, CV):
         """Initializes the CV window."""
-        CA.setObjectName("CA")
-        CA.resize(808, 476)
+        CV.setObjectName("CV")
+        CV.resize(871, 596)
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        CV.setPalette(palette)
         icon = QtGui.QIcon()
         icon.addPixmap(
             QtGui.QPixmap("../pics/icon_pytentiostat.ico"),
             QtGui.QIcon.Normal,
             QtGui.QIcon.Off,
         )
-        CA.setWindowIcon(icon)
-        self.centralwidget = QtWidgets.QWidget(CA)
+        CV.setWindowIcon(icon)
+        self.centralwidget = QtWidgets.QWidget(CV)
         self.centralwidget.setObjectName("centralwidget")
         self.rest_time_label = QtWidgets.QLabel(self.centralwidget)
         self.rest_time_label.setEnabled(True)
-        self.rest_time_label.setGeometry(QtCore.QRect(10, 220, 161, 31))
+        self.rest_time_label.setGeometry(QtCore.QRect(10, 230, 161, 31))
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
         brush.setStyle(QtCore.Qt.SolidPattern)
@@ -443,29 +464,6 @@ class Ui_CA(object):
             QtCore.Qt.LinksAccessibleByMouse
         )
         self.experiment_type_label.setObjectName("experiment_type_label")
-        self.experiment_step_number = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_step_number.setGeometry(QtCore.QRect(180, 260, 61, 31))
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.experiment_step_number.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.experiment_step_number.setFont(font)
-        self.experiment_step_number.setStyleSheet("")
-        self.experiment_step_number.setText("")
-        self.experiment_step_number.setFrame(True)
-        self.experiment_step_number.setAlignment(QtCore.Qt.AlignCenter)
-        self.experiment_step_number.setObjectName("experiment_step_number")
         self.select_output_filepath_button = QtWidgets.QPushButton(
             self.centralwidget
         )
@@ -645,9 +643,9 @@ class Ui_CA(object):
         self.select_output_filepath_button.setObjectName(
             "select_output_filepath_button"
         )
-        self.experiment_time_label = QtWidgets.QLabel(self.centralwidget)
-        self.experiment_time_label.setEnabled(True)
-        self.experiment_time_label.setGeometry(QtCore.QRect(10, 340, 161, 31))
+        self.sweep_rate_label = QtWidgets.QLabel(self.centralwidget)
+        self.sweep_rate_label.setEnabled(True)
+        self.sweep_rate_label.setGeometry(QtCore.QRect(10, 470, 161, 31))
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
         brush.setStyle(QtCore.Qt.SolidPattern)
@@ -824,26 +822,227 @@ class Ui_CA(object):
         palette.setBrush(
             QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipText, brush
         )
-        self.experiment_time_label.setPalette(palette)
+        self.sweep_rate_label.setPalette(palette)
         font = QtGui.QFont()
         font.setFamily("Arial")
         font.setBold(True)
         font.setWeight(font.Weight.Medium)
-        self.experiment_time_label.setFont(font)
-        self.experiment_time_label.setContextMenuPolicy(
-            QtCore.Qt.NoContextMenu
-        )
-        self.experiment_time_label.setAcceptDrops(False)
-        self.experiment_time_label.setAutoFillBackground(True)
-        self.experiment_time_label.setFrameShape(QtWidgets.QFrame.Box)
-        self.experiment_time_label.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.experiment_time_label.setLineWidth(1)
-        self.experiment_time_label.setMidLineWidth(1)
-        self.experiment_time_label.setScaledContents(False)
-        self.experiment_time_label.setTextInteractionFlags(
+        self.sweep_rate_label.setFont(font)
+        self.sweep_rate_label.setContextMenuPolicy(QtCore.Qt.NoContextMenu)
+        self.sweep_rate_label.setAcceptDrops(False)
+        self.sweep_rate_label.setAutoFillBackground(True)
+        self.sweep_rate_label.setFrameShape(QtWidgets.QFrame.Box)
+        self.sweep_rate_label.setFrameShadow(QtWidgets.QFrame.Sunken)
+        self.sweep_rate_label.setLineWidth(1)
+        self.sweep_rate_label.setMidLineWidth(1)
+        self.sweep_rate_label.setScaledContents(False)
+        self.sweep_rate_label.setTextInteractionFlags(
             QtCore.Qt.LinksAccessibleByMouse
         )
-        self.experiment_time_label.setObjectName("experiment_time_label")
+        self.sweep_rate_label.setObjectName("sweep_rate_label")
+        self.first_vertex_voltage_label = QtWidgets.QLabel(self.centralwidget)
+        self.first_vertex_voltage_label.setEnabled(True)
+        self.first_vertex_voltage_label.setGeometry(
+            QtCore.QRect(10, 390, 161, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipText, brush
+        )
+        self.first_vertex_voltage_label.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.first_vertex_voltage_label.setFont(font)
+        self.first_vertex_voltage_label.setContextMenuPolicy(
+            QtCore.Qt.NoContextMenu
+        )
+        self.first_vertex_voltage_label.setAcceptDrops(False)
+        self.first_vertex_voltage_label.setAutoFillBackground(True)
+        self.first_vertex_voltage_label.setFrameShape(QtWidgets.QFrame.Box)
+        self.first_vertex_voltage_label.setFrameShadow(QtWidgets.QFrame.Sunken)
+        self.first_vertex_voltage_label.setLineWidth(1)
+        self.first_vertex_voltage_label.setMidLineWidth(1)
+        self.first_vertex_voltage_label.setScaledContents(False)
+        self.first_vertex_voltage_label.setTextInteractionFlags(
+            QtCore.Qt.LinksAccessibleByMouse
+        )
+        self.first_vertex_voltage_label.setObjectName(
+            "first_vertex_voltage_label"
+        )
         self.general_parameters_label = QtWidgets.QLabel(self.centralwidget)
         self.general_parameters_label.setEnabled(True)
         self.general_parameters_label.setGeometry(
@@ -1045,31 +1244,8 @@ class Ui_CA(object):
             QtCore.Qt.LinksAccessibleByMouse
         )
         self.general_parameters_label.setObjectName("general_parameters_label")
-        self.experiment_time = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_time.setGeometry(QtCore.QRect(180, 340, 61, 31))
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.experiment_time.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.experiment_time.setFont(font)
-        self.experiment_time.setStyleSheet("")
-        self.experiment_time.setText("")
-        self.experiment_time.setFrame(True)
-        self.experiment_time.setAlignment(QtCore.Qt.AlignCenter)
-        self.experiment_time.setObjectName("experiment_time")
         self.experiment_type_verify = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_type_verify.setGeometry(QtCore.QRect(180, 50, 241, 31))
+        self.experiment_type_verify.setGeometry(QtCore.QRect(180, 50, 251, 31))
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
         brush.setStyle(QtCore.Qt.SolidPattern)
@@ -1120,29 +1296,6 @@ class Ui_CA(object):
         self.experiment_type_verify.setAlignment(QtCore.Qt.AlignCenter)
         self.experiment_type_verify.setReadOnly(True)
         self.experiment_type_verify.setObjectName("experiment_type_verify")
-        self.experiment_voltage = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_voltage.setGeometry(QtCore.QRect(180, 300, 61, 31))
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.experiment_voltage.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.experiment_voltage.setFont(font)
-        self.experiment_voltage.setStyleSheet("")
-        self.experiment_voltage.setText("")
-        self.experiment_voltage.setFrame(True)
-        self.experiment_voltage.setAlignment(QtCore.Qt.AlignCenter)
-        self.experiment_voltage.setObjectName("experiment_voltage")
         self.output_filename_label = QtWidgets.QLabel(self.centralwidget)
         self.output_filename_label.setEnabled(True)
         self.output_filename_label.setGeometry(QtCore.QRect(10, 90, 161, 31))
@@ -1343,7 +1496,7 @@ class Ui_CA(object):
         )
         self.output_filename_label.setObjectName("output_filename_label")
         self.experiment_file_name = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_file_name.setGeometry(QtCore.QRect(180, 90, 241, 31))
+        self.experiment_file_name.setGeometry(QtCore.QRect(180, 90, 251, 31))
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
         brush.setStyle(QtCore.Qt.SolidPattern)
@@ -1368,7 +1521,7 @@ class Ui_CA(object):
         self.experiment_file_name.setObjectName("experiment_file_name")
         self.step_number_label = QtWidgets.QLabel(self.centralwidget)
         self.step_number_label.setEnabled(True)
-        self.step_number_label.setGeometry(QtCore.QRect(10, 260, 161, 31))
+        self.step_number_label.setGeometry(QtCore.QRect(10, 270, 161, 31))
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
         brush.setStyle(QtCore.Qt.SolidPattern)
@@ -1563,6 +1716,29 @@ class Ui_CA(object):
             QtCore.Qt.LinksAccessibleByMouse
         )
         self.step_number_label.setObjectName("step_number_label")
+        self.experiment_sweep_rate = QtWidgets.QLineEdit(self.centralwidget)
+        self.experiment_sweep_rate.setGeometry(QtCore.QRect(180, 470, 61, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.experiment_sweep_rate.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.experiment_sweep_rate.setFont(font)
+        self.experiment_sweep_rate.setStyleSheet("")
+        self.experiment_sweep_rate.setText("")
+        self.experiment_sweep_rate.setFrame(True)
+        self.experiment_sweep_rate.setAlignment(QtCore.Qt.AlignCenter)
+        self.experiment_sweep_rate.setObjectName("experiment_sweep_rate")
         self.experiment_parameters_label = QtWidgets.QLabel(self.centralwidget)
         self.experiment_parameters_label.setEnabled(True)
         self.experiment_parameters_label.setGeometry(
@@ -1768,32 +1944,8 @@ class Ui_CA(object):
         self.experiment_parameters_label.setObjectName(
             "experiment_parameters_label"
         )
-        self.experiment_duration = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_duration.setGeometry(QtCore.QRect(220, 390, 181, 31))
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.experiment_duration.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.experiment_duration.setFont(font)
-        self.experiment_duration.setStyleSheet("")
-        self.experiment_duration.setText("")
-        self.experiment_duration.setFrame(True)
-        self.experiment_duration.setAlignment(QtCore.Qt.AlignCenter)
-        self.experiment_duration.setReadOnly(True)
-        self.experiment_duration.setObjectName("experiment_duration")
         self.experiment_file_path = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_file_path.setGeometry(QtCore.QRect(180, 130, 241, 31))
+        self.experiment_file_path.setGeometry(QtCore.QRect(180, 130, 251, 31))
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
         brush.setStyle(QtCore.Qt.SolidPattern)
@@ -1820,7 +1972,7 @@ class Ui_CA(object):
         self.experiment_preview_label = QtWidgets.QLabel(self.centralwidget)
         self.experiment_preview_label.setEnabled(True)
         self.experiment_preview_label.setGeometry(
-            QtCore.QRect(430, 10, 351, 31)
+            QtCore.QRect(440, 10, 421, 31)
         )
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
@@ -2018,10 +2170,1643 @@ class Ui_CA(object):
             QtCore.Qt.LinksAccessibleByMouse
         )
         self.experiment_preview_label.setObjectName("experiment_preview_label")
+        self.sweep_rate_units_label = QtWidgets.QLineEdit(self.centralwidget)
+        self.sweep_rate_units_label.setGeometry(QtCore.QRect(250, 470, 41, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.sweep_rate_units_label.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.sweep_rate_units_label.setFont(font)
+        self.sweep_rate_units_label.setStatusTip("")
+        self.sweep_rate_units_label.setStyleSheet(
+            "background-color: rgb(240, 240, 240);"
+        )
+        self.sweep_rate_units_label.setFrame(False)
+        self.sweep_rate_units_label.setObjectName("sweep_rate_units_label")
+        self.v_vs_label_1 = QtWidgets.QLineEdit(self.centralwidget)
+        self.v_vs_label_1.setGeometry(QtCore.QRect(250, 350, 41, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.v_vs_label_1.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.v_vs_label_1.setFont(font)
+        self.v_vs_label_1.setStatusTip("")
+        self.v_vs_label_1.setStyleSheet(
+            "background-color: rgb(240, 240, 240);"
+        )
+        self.v_vs_label_1.setFrame(False)
+        self.v_vs_label_1.setObjectName("v_vs_label_1")
+        self.v_vs_label_2 = QtWidgets.QLineEdit(self.centralwidget)
+        self.v_vs_label_2.setGeometry(QtCore.QRect(250, 390, 41, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.v_vs_label_2.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.v_vs_label_2.setFont(font)
+        self.v_vs_label_2.setStatusTip("")
+        self.v_vs_label_2.setStyleSheet(
+            "background-color: rgb(240, 240, 240);"
+        )
+        self.v_vs_label_2.setFrame(False)
+        self.v_vs_label_2.setObjectName("v_vs_label_2")
+        self.experiment_rest_time = QtWidgets.QLineEdit(self.centralwidget)
+        self.experiment_rest_time.setGeometry(QtCore.QRect(180, 230, 61, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.experiment_rest_time.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.experiment_rest_time.setFont(font)
+        self.experiment_rest_time.setStyleSheet("")
+        self.experiment_rest_time.setText("")
+        self.experiment_rest_time.setFrame(True)
+        self.experiment_rest_time.setAlignment(QtCore.Qt.AlignCenter)
+        self.experiment_rest_time.setObjectName("experiment_rest_time")
+        self.advanced_parameters_button = QtWidgets.QPushButton(
+            self.centralwidget
+        )
+        self.advanced_parameters_button.setGeometry(
+            QtCore.QRect(440, 510, 211, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(91, 166, 232))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.NoRole, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(248, 221, 23))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.NoRole, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 120, 215))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.NoRole, brush)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.advanced_parameters_button.setFont(font)
+        self.advanced_parameters_button.setContextMenuPolicy(
+            QtCore.Qt.PreventContextMenu
+        )
+        self.advanced_parameters_button.setAcceptDrops(False)
+        self.advanced_parameters_button.setWhatsThis("")
+        self.advanced_parameters_button.setAutoFillBackground(True)
+        self.advanced_parameters_button.setLocale(
+            QtCore.QLocale(QtCore.QLocale.English, QtCore.QLocale.UnitedStates)
+        )
+        self.advanced_parameters_button.setInputMethodHints(QtCore.Qt.ImhNone)
+        self.advanced_parameters_button.setAutoRepeatDelay(301)
+        self.advanced_parameters_button.setAutoRepeatInterval(96)
+        self.advanced_parameters_button.setAutoDefault(False)
+        self.advanced_parameters_button.setDefault(False)
+        self.advanced_parameters_button.setFlat(False)
+        self.advanced_parameters_button.setObjectName(
+            "advanced_parameters_button"
+        )
+        self.start_voltage_label = QtWidgets.QLabel(self.centralwidget)
+        self.start_voltage_label.setEnabled(True)
+        self.start_voltage_label.setGeometry(QtCore.QRect(10, 350, 161, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipText, brush
+        )
+        self.start_voltage_label.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.start_voltage_label.setFont(font)
+        self.start_voltage_label.setContextMenuPolicy(QtCore.Qt.NoContextMenu)
+        self.start_voltage_label.setAcceptDrops(False)
+        self.start_voltage_label.setAutoFillBackground(True)
+        self.start_voltage_label.setFrameShape(QtWidgets.QFrame.Box)
+        self.start_voltage_label.setFrameShadow(QtWidgets.QFrame.Sunken)
+        self.start_voltage_label.setLineWidth(1)
+        self.start_voltage_label.setMidLineWidth(1)
+        self.start_voltage_label.setScaledContents(False)
+        self.start_voltage_label.setTextInteractionFlags(
+            QtCore.Qt.LinksAccessibleByMouse
+        )
+        self.start_voltage_label.setObjectName("start_voltage_label")
+        self.rest_time_units_label = QtWidgets.QLineEdit(self.centralwidget)
+        self.rest_time_units_label.setGeometry(QtCore.QRect(250, 230, 16, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.rest_time_units_label.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.rest_time_units_label.setFont(font)
+        self.rest_time_units_label.setStatusTip("")
+        self.rest_time_units_label.setStyleSheet(
+            "background-color: rgb(240, 240, 240);"
+        )
+        self.rest_time_units_label.setFrame(False)
+        self.rest_time_units_label.setObjectName("rest_time_units_label")
+        self.save_experiment_file_button = QtWidgets.QPushButton(
+            self.centralwidget
+        )
+        self.save_experiment_file_button.setGeometry(
+            QtCore.QRect(660, 510, 201, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(91, 166, 232))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.NoRole, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(248, 221, 23))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.NoRole, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 120, 215))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.NoRole, brush)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.save_experiment_file_button.setFont(font)
+        self.save_experiment_file_button.setContextMenuPolicy(
+            QtCore.Qt.PreventContextMenu
+        )
+        self.save_experiment_file_button.setAcceptDrops(False)
+        self.save_experiment_file_button.setWhatsThis("")
+        self.save_experiment_file_button.setAutoFillBackground(True)
+        self.save_experiment_file_button.setLocale(
+            QtCore.QLocale(QtCore.QLocale.English, QtCore.QLocale.UnitedStates)
+        )
+        self.save_experiment_file_button.setInputMethodHints(QtCore.Qt.ImhNone)
+        self.save_experiment_file_button.setAutoRepeatDelay(301)
+        self.save_experiment_file_button.setAutoRepeatInterval(96)
+        self.save_experiment_file_button.setAutoDefault(False)
+        self.save_experiment_file_button.setDefault(False)
+        self.save_experiment_file_button.setFlat(False)
+        self.save_experiment_file_button.setObjectName(
+            "save_experiment_file_button"
+        )
+        self.generate_preview_button = QtWidgets.QPushButton(
+            self.centralwidget
+        )
+        self.generate_preview_button.setGeometry(
+            QtCore.QRect(660, 410, 201, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(91, 166, 232))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.NoRole, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(248, 221, 23))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.NoRole, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 120, 215))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Highlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Link, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.NoRole, brush)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.generate_preview_button.setFont(font)
+        self.generate_preview_button.setContextMenuPolicy(
+            QtCore.Qt.PreventContextMenu
+        )
+        self.generate_preview_button.setAcceptDrops(False)
+        self.generate_preview_button.setWhatsThis("")
+        self.generate_preview_button.setAutoFillBackground(True)
+        self.generate_preview_button.setLocale(
+            QtCore.QLocale(QtCore.QLocale.English, QtCore.QLocale.UnitedStates)
+        )
+        self.generate_preview_button.setInputMethodHints(QtCore.Qt.ImhNone)
+        self.generate_preview_button.setAutoRepeatDelay(301)
+        self.generate_preview_button.setAutoRepeatInterval(96)
+        self.generate_preview_button.setAutoDefault(False)
+        self.generate_preview_button.setDefault(False)
+        self.generate_preview_button.setFlat(False)
+        self.generate_preview_button.setObjectName("generate_preview_button")
+        self.second_vertex_voltage_label = QtWidgets.QLabel(self.centralwidget)
+        self.second_vertex_voltage_label.setEnabled(True)
+        self.second_vertex_voltage_label.setGeometry(
+            QtCore.QRect(10, 430, 161, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipText, brush
+        )
+        self.second_vertex_voltage_label.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.second_vertex_voltage_label.setFont(font)
+        self.second_vertex_voltage_label.setContextMenuPolicy(
+            QtCore.Qt.NoContextMenu
+        )
+        self.second_vertex_voltage_label.setAcceptDrops(False)
+        self.second_vertex_voltage_label.setAutoFillBackground(True)
+        self.second_vertex_voltage_label.setFrameShape(QtWidgets.QFrame.Box)
+        self.second_vertex_voltage_label.setFrameShadow(
+            QtWidgets.QFrame.Sunken
+        )
+        self.second_vertex_voltage_label.setLineWidth(1)
+        self.second_vertex_voltage_label.setMidLineWidth(1)
+        self.second_vertex_voltage_label.setScaledContents(False)
+        self.second_vertex_voltage_label.setTextInteractionFlags(
+            QtCore.Qt.LinksAccessibleByMouse
+        )
+        self.second_vertex_voltage_label.setObjectName(
+            "second_vertex_voltage_label"
+        )
+        self.experiment_second_vertex_voltage = QtWidgets.QLineEdit(
+            self.centralwidget
+        )
+        self.experiment_second_vertex_voltage.setGeometry(
+            QtCore.QRect(180, 430, 61, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.experiment_second_vertex_voltage.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.experiment_second_vertex_voltage.setFont(font)
+        self.experiment_second_vertex_voltage.setStyleSheet("")
+        self.experiment_second_vertex_voltage.setText("")
+        self.experiment_second_vertex_voltage.setFrame(True)
+        self.experiment_second_vertex_voltage.setAlignment(
+            QtCore.Qt.AlignCenter
+        )
+        self.experiment_second_vertex_voltage.setObjectName(
+            "experiment_second_vertex_voltage"
+        )
+        self.v_vs_label_3 = QtWidgets.QLineEdit(self.centralwidget)
+        self.v_vs_label_3.setGeometry(QtCore.QRect(250, 430, 41, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.v_vs_label_3.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.v_vs_label_3.setFont(font)
+        self.v_vs_label_3.setStatusTip("")
+        self.v_vs_label_3.setStyleSheet(
+            "background-color: rgb(240, 240, 240);"
+        )
+        self.v_vs_label_3.setFrame(False)
+        self.v_vs_label_3.setObjectName("v_vs_label_3")
+        self.cycle_number_label = QtWidgets.QLabel(self.centralwidget)
+        self.cycle_number_label.setEnabled(True)
+        self.cycle_number_label.setGeometry(QtCore.QRect(10, 310, 161, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Active, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
+        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
+        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.BrightText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipBase, brush
+        )
+        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(
+            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipText, brush
+        )
+        self.cycle_number_label.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.cycle_number_label.setFont(font)
+        self.cycle_number_label.setContextMenuPolicy(QtCore.Qt.NoContextMenu)
+        self.cycle_number_label.setAcceptDrops(False)
+        self.cycle_number_label.setAutoFillBackground(True)
+        self.cycle_number_label.setFrameShape(QtWidgets.QFrame.Box)
+        self.cycle_number_label.setFrameShadow(QtWidgets.QFrame.Sunken)
+        self.cycle_number_label.setLineWidth(1)
+        self.cycle_number_label.setMidLineWidth(1)
+        self.cycle_number_label.setScaledContents(False)
+        self.cycle_number_label.setTextInteractionFlags(
+            QtCore.Qt.LinksAccessibleByMouse
+        )
+        self.cycle_number_label.setObjectName("cycle_number_label")
+        self.experiment_cycle_number = QtWidgets.QLineEdit(self.centralwidget)
+        self.experiment_cycle_number.setGeometry(
+            QtCore.QRect(180, 310, 61, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.experiment_cycle_number.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.experiment_cycle_number.setFont(font)
+        self.experiment_cycle_number.setStyleSheet("")
+        self.experiment_cycle_number.setText("")
+        self.experiment_cycle_number.setFrame(True)
+        self.experiment_cycle_number.setAlignment(QtCore.Qt.AlignCenter)
+        self.experiment_cycle_number.setObjectName("experiment_cycle_number")
+        self.experiment_step_number = QtWidgets.QLineEdit(self.centralwidget)
+        self.experiment_step_number.setGeometry(QtCore.QRect(180, 270, 61, 31))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.experiment_step_number.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.experiment_step_number.setFont(font)
+        self.experiment_step_number.setStyleSheet("")
+        self.experiment_step_number.setText("")
+        self.experiment_step_number.setFrame(True)
+        self.experiment_step_number.setAlignment(QtCore.Qt.AlignCenter)
+        self.experiment_step_number.setObjectName("experiment_step_number")
+        self.experiment_start_voltage = QtWidgets.QLineEdit(self.centralwidget)
+        self.experiment_start_voltage.setGeometry(
+            QtCore.QRect(180, 350, 61, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.experiment_start_voltage.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.experiment_start_voltage.setFont(font)
+        self.experiment_start_voltage.setStyleSheet("")
+        self.experiment_start_voltage.setText("")
+        self.experiment_start_voltage.setFrame(True)
+        self.experiment_start_voltage.setAlignment(QtCore.Qt.AlignCenter)
+        self.experiment_start_voltage.setObjectName("experiment_start_voltage")
+        self.experiment_first_vertex_voltage = QtWidgets.QLineEdit(
+            self.centralwidget
+        )
+        self.experiment_first_vertex_voltage.setGeometry(
+            QtCore.QRect(180, 390, 61, 31)
+        )
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
+        self.experiment_first_vertex_voltage.setPalette(palette)
+        font = QtGui.QFont()
+        font.setFamily("Arial")
+        font.setBold(True)
+        font.setWeight(font.Weight.Medium)
+        self.experiment_first_vertex_voltage.setFont(font)
+        self.experiment_first_vertex_voltage.setStyleSheet("")
+        self.experiment_first_vertex_voltage.setText("")
+        self.experiment_first_vertex_voltage.setFrame(True)
+        self.experiment_first_vertex_voltage.setAlignment(
+            QtCore.Qt.AlignCenter
+        )
+        self.experiment_first_vertex_voltage.setObjectName(
+            "experiment_first_vertex_voltage"
+        )
+        self.voltage_ref = QtWidgets.QComboBox(self.centralwidget)
+        self.voltage_ref.setGeometry(QtCore.QRect(300, 350, 101, 31))
+        font = QtGui.QFont()
+        font.setBold(False)
+        font.setWeight(font.Weight.Medium)
+        self.voltage_ref.setFont(font)
+        self.voltage_ref.setEditable(False)
+        self.voltage_ref.setInsertPolicy(QtWidgets.QComboBox.InsertAtBottom)
+        self.voltage_ref.setObjectName("voltage_ref")
+        self.voltage_ref.addItem("")
+        self.voltage_ref.addItem("")
+        self.voltage_ref_2 = QtWidgets.QComboBox(self.centralwidget)
+        self.voltage_ref_2.setGeometry(QtCore.QRect(300, 390, 101, 31))
+        font = QtGui.QFont()
+        font.setBold(False)
+        font.setWeight(font.Weight.Medium)
+        self.voltage_ref_2.setFont(font)
+        self.voltage_ref_2.setEditable(False)
+        self.voltage_ref_2.setInsertPolicy(QtWidgets.QComboBox.InsertAtBottom)
+        self.voltage_ref_2.setObjectName("voltage_ref_2")
+        self.voltage_ref_2.addItem("")
+        self.voltage_ref_2.addItem("")
+        self.voltage_ref_3 = QtWidgets.QComboBox(self.centralwidget)
+        self.voltage_ref_3.setGeometry(QtCore.QRect(300, 430, 101, 31))
+        font = QtGui.QFont()
+        font.setBold(False)
+        font.setWeight(font.Weight.Medium)
+        self.voltage_ref_3.setFont(font)
+        self.voltage_ref_3.setEditable(False)
+        self.voltage_ref_3.setInsertPolicy(QtWidgets.QComboBox.InsertAtBottom)
+        self.voltage_ref_3.setObjectName("voltage_ref_3")
+        self.voltage_ref_3.addItem("")
+        self.voltage_ref_3.addItem("")
+        self.plot_area = QtWidgets.QWidget(self.centralwidget)
+        self.plot_area.setGeometry(QtCore.QRect(440, 50, 421, 337))
+        palette = QtGui.QPalette()
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
+        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
+        brush.setStyle(QtCore.Qt.SolidPattern)
+        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
+        self.plot_area.setPalette(palette)
+        self.plot_area.setStyleSheet(
+            "background-color: rgb(255, 255, 255);\n"
+            "border: 1px solid black;\n"
+            "\n"
+            ""
+        )
+        self.plot_area.setObjectName("plot_area")
         self.experiment_duration_button = QtWidgets.QLabel(self.centralwidget)
         self.experiment_duration_button.setEnabled(True)
         self.experiment_duration_button.setGeometry(
-            QtCore.QRect(10, 390, 201, 31)
+            QtCore.QRect(10, 510, 191, 31)
         )
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
@@ -2221,57 +4006,8 @@ class Ui_CA(object):
         self.experiment_duration_button.setObjectName(
             "experiment_duration_button"
         )
-        self.v_vs_label = QtWidgets.QLineEdit(self.centralwidget)
-        self.v_vs_label.setGeometry(QtCore.QRect(250, 300, 51, 31))
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.v_vs_label.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.v_vs_label.setFont(font)
-        self.v_vs_label.setStatusTip("")
-        self.v_vs_label.setStyleSheet("background-color: rgb(240, 240, 240);")
-        self.v_vs_label.setFrame(False)
-        self.v_vs_label.setObjectName("v_vs_label")
-        self.experiment_rest_time = QtWidgets.QLineEdit(self.centralwidget)
-        self.experiment_rest_time.setGeometry(QtCore.QRect(180, 220, 61, 31))
+        self.experiment_duration = QtWidgets.QLineEdit(self.centralwidget)
+        self.experiment_duration.setGeometry(QtCore.QRect(210, 510, 221, 31))
         palette = QtGui.QPalette()
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
         brush.setStyle(QtCore.Qt.SolidPattern)
@@ -2282,1059 +4018,215 @@ class Ui_CA(object):
         brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
         brush.setStyle(QtCore.Qt.SolidPattern)
         palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.experiment_rest_time.setPalette(palette)
+        self.experiment_duration.setPalette(palette)
         font = QtGui.QFont()
         font.setFamily("Arial")
         font.setBold(True)
         font.setWeight(font.Weight.Medium)
-        self.experiment_rest_time.setFont(font)
-        self.experiment_rest_time.setStyleSheet("")
-        self.experiment_rest_time.setText("")
-        self.experiment_rest_time.setFrame(True)
-        self.experiment_rest_time.setAlignment(QtCore.Qt.AlignCenter)
-        self.experiment_rest_time.setObjectName("experiment_rest_time")
-        self.advanced_parameters_button = QtWidgets.QPushButton(
-            self.centralwidget
-        )
-        self.advanced_parameters_button.setGeometry(
-            QtCore.QRect(410, 390, 191, 31)
-        )
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(91, 166, 232))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.NoRole, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(248, 221, 23))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.NoRole, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 120, 215))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.NoRole, brush)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.advanced_parameters_button.setFont(font)
-        self.advanced_parameters_button.setContextMenuPolicy(
-            QtCore.Qt.PreventContextMenu
-        )
-        self.advanced_parameters_button.setAcceptDrops(False)
-        self.advanced_parameters_button.setWhatsThis("")
-        self.advanced_parameters_button.setAutoFillBackground(True)
-        self.advanced_parameters_button.setLocale(
-            QtCore.QLocale(QtCore.QLocale.English, QtCore.QLocale.UnitedStates)
-        )
-        self.advanced_parameters_button.setInputMethodHints(QtCore.Qt.ImhNone)
-        self.advanced_parameters_button.setAutoRepeatDelay(301)
-        self.advanced_parameters_button.setAutoRepeatInterval(96)
-        self.advanced_parameters_button.setAutoDefault(False)
-        self.advanced_parameters_button.setDefault(False)
-        self.advanced_parameters_button.setFlat(False)
-        self.advanced_parameters_button.setObjectName(
-            "advanced_parameters_button"
-        )
-        self.voltage_label = QtWidgets.QLabel(self.centralwidget)
-        self.voltage_label.setEnabled(True)
-        self.voltage_label.setGeometry(QtCore.QRect(10, 300, 161, 31))
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
-        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.BrightText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.ToolTipBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.ToolTipText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.BrightText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(161, 188, 213))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.ToolTipText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(84, 151, 213))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(44, 80, 114))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.BrightText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(33, 60, 85))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(67, 121, 171))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 220))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.ToolTipText, brush
-        )
-        self.voltage_label.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.voltage_label.setFont(font)
-        self.voltage_label.setContextMenuPolicy(QtCore.Qt.NoContextMenu)
-        self.voltage_label.setAcceptDrops(False)
-        self.voltage_label.setAutoFillBackground(True)
-        self.voltage_label.setFrameShape(QtWidgets.QFrame.Box)
-        self.voltage_label.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.voltage_label.setLineWidth(1)
-        self.voltage_label.setMidLineWidth(1)
-        self.voltage_label.setScaledContents(False)
-        self.voltage_label.setTextInteractionFlags(
-            QtCore.Qt.LinksAccessibleByMouse
-        )
-        self.voltage_label.setObjectName("voltage_label")
-        self.rest_time_units_label = QtWidgets.QLineEdit(self.centralwidget)
-        self.rest_time_units_label.setGeometry(QtCore.QRect(250, 220, 16, 31))
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.rest_time_units_label.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.rest_time_units_label.setFont(font)
-        self.rest_time_units_label.setStatusTip("")
-        self.rest_time_units_label.setStyleSheet(
-            "background-color: rgb(240, 240, 240);"
-        )
-        self.rest_time_units_label.setFrame(False)
-        self.rest_time_units_label.setObjectName("rest_time_units_label")
-        self.save_experiment_file_button = QtWidgets.QPushButton(
-            self.centralwidget
-        )
-        self.save_experiment_file_button.setGeometry(
-            QtCore.QRect(610, 390, 171, 31)
-        )
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(91, 166, 232))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.NoRole, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(248, 221, 23))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.NoRole, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 120, 215))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.NoRole, brush)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.save_experiment_file_button.setFont(font)
-        self.save_experiment_file_button.setContextMenuPolicy(
-            QtCore.Qt.PreventContextMenu
-        )
-        self.save_experiment_file_button.setAcceptDrops(False)
-        self.save_experiment_file_button.setWhatsThis("")
-        self.save_experiment_file_button.setAutoFillBackground(True)
-        self.save_experiment_file_button.setLocale(
-            QtCore.QLocale(QtCore.QLocale.English, QtCore.QLocale.UnitedStates)
-        )
-        self.save_experiment_file_button.setInputMethodHints(QtCore.Qt.ImhNone)
-        self.save_experiment_file_button.setAutoRepeatDelay(301)
-        self.save_experiment_file_button.setAutoRepeatInterval(96)
-        self.save_experiment_file_button.setAutoDefault(False)
-        self.save_experiment_file_button.setDefault(False)
-        self.save_experiment_file_button.setFlat(False)
-        self.save_experiment_file_button.setObjectName(
-            "save_experiment_file_button"
-        )
-        self.generate_preview_button = QtWidgets.QPushButton(
-            self.centralwidget
-        )
-        self.generate_preview_button.setGeometry(
-            QtCore.QRect(610, 350, 171, 31)
-        )
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Midlight, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(91, 166, 232))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Active, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.NoRole, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(248, 221, 23))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Inactive, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.NoRole, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.WindowText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Light, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.Midlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Dark, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Mid, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Text, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.ButtonText, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 120, 215))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.Highlight, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(247, 217, 21))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Link, brush)
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(
-            QtGui.QPalette.Disabled, QtGui.QPalette.AlternateBase, brush
-        )
-        brush = QtGui.QBrush(QtGui.QColor(0, 0, 0))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.NoRole, brush)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.generate_preview_button.setFont(font)
-        self.generate_preview_button.setContextMenuPolicy(
-            QtCore.Qt.PreventContextMenu
-        )
-        self.generate_preview_button.setAcceptDrops(False)
-        self.generate_preview_button.setWhatsThis("")
-        self.generate_preview_button.setAutoFillBackground(True)
-        self.generate_preview_button.setLocale(
-            QtCore.QLocale(QtCore.QLocale.English, QtCore.QLocale.UnitedStates)
-        )
-        self.generate_preview_button.setInputMethodHints(QtCore.Qt.ImhNone)
-        self.generate_preview_button.setAutoRepeatDelay(301)
-        self.generate_preview_button.setAutoRepeatInterval(96)
-        self.generate_preview_button.setAutoDefault(False)
-        self.generate_preview_button.setDefault(False)
-        self.generate_preview_button.setFlat(False)
-        self.generate_preview_button.setObjectName("generate_preview_button")
-        self.plot_area = QtWidgets.QWidget(self.centralwidget)
-        self.plot_area.setGeometry(QtCore.QRect(430, 50, 351, 281))
-        self.plot_area.setStyleSheet(
-            "border: 1px solid black;\n"
-            "background-color: rgb(255, 255, 255);\n"
-            "\n"
-            ""
-        )
-        self.plot_area.setObjectName("plot_area")
-        self.rest_time_units_label_2 = QtWidgets.QLineEdit(self.centralwidget)
-        self.rest_time_units_label_2.setGeometry(
-            QtCore.QRect(250, 340, 16, 31)
-        )
-        palette = QtGui.QPalette()
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Active, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Inactive, QtGui.QPalette.Shadow, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Button, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Base, brush)
-        brush = QtGui.QBrush(QtGui.QColor(240, 240, 240))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Window, brush)
-        brush = QtGui.QBrush(QtGui.QColor(255, 255, 255))
-        brush.setStyle(QtCore.Qt.SolidPattern)
-        palette.setBrush(QtGui.QPalette.Disabled, QtGui.QPalette.Shadow, brush)
-        self.rest_time_units_label_2.setPalette(palette)
-        font = QtGui.QFont()
-        font.setFamily("Arial")
-        font.setBold(True)
-        font.setWeight(font.Weight.Medium)
-        self.rest_time_units_label_2.setFont(font)
-        self.rest_time_units_label_2.setStatusTip("")
-        self.rest_time_units_label_2.setStyleSheet(
-            "\n" "background-color: rgb(240, 240, 240);"
-        )
-        self.rest_time_units_label_2.setFrame(False)
-        self.rest_time_units_label_2.setObjectName("rest_time_units_label_2")
-        self.voltage_ref = QtWidgets.QComboBox(self.centralwidget)
-        self.voltage_ref.setGeometry(QtCore.QRect(320, 300, 101, 31))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(font.Weight.Medium)
-        self.voltage_ref.setFont(font)
-        self.voltage_ref.setEditable(False)
-        self.voltage_ref.setInsertPolicy(QtWidgets.QComboBox.InsertAtBottom)
-        self.voltage_ref.setObjectName("voltage_ref")
-        self.voltage_ref.addItem("")
-        self.voltage_ref.addItem("")
-        CA.setCentralWidget(self.centralwidget)
-        self.menubar = QtWidgets.QMenuBar(CA)
-        self.menubar.setGeometry(QtCore.QRect(0, 0, 808, 18))
+        self.experiment_duration.setFont(font)
+        self.experiment_duration.setStyleSheet("")
+        self.experiment_duration.setText("")
+        self.experiment_duration.setFrame(True)
+        self.experiment_duration.setAlignment(QtCore.Qt.AlignCenter)
+        self.experiment_duration.setReadOnly(True)
+        self.experiment_duration.setObjectName("experiment_duration")
+        CV.setCentralWidget(self.centralwidget)
+        self.menubar = QtWidgets.QMenuBar(CV)
+        self.menubar.setGeometry(QtCore.QRect(0, 0, 871, 18))
         self.menubar.setObjectName("menubar")
-        CA.setMenuBar(self.menubar)
-        self.statusbar = QtWidgets.QStatusBar(CA)
-        self.statusbar.setStyleSheet(
-            "background-color: rgb(236, 236, 236);\n"
-            "background-color: rgb(228, 228, 228);"
-        )
+        CV.setMenuBar(self.menubar)
+        self.statusbar = QtWidgets.QStatusBar(CV)
+        self.statusbar.setStyleSheet("background-color: rgb(228, 228, 228);")
         self.statusbar.setObjectName("statusbar")
-        CA.setStatusBar(self.statusbar)
-        self.actionVref = QtGui.QAction(CA)
-        self.actionVref.setObjectName("actionVref")
-        self.actionVocv = QtGui.QAction(CA)
-        self.actionVocv.setObjectName("actionVocv")
+        CV.setStatusBar(self.statusbar)
 
-        self.retranslateUi(CA)
-        QtCore.QMetaObject.connectSlotsByName(CA)
+        self.retranslateUi(CV)
+        QtCore.QMetaObject.connectSlotsByName(CV)
 
-    def retranslateUi(self, CA):
+    def retranslateUi(self, CV):
         _translate = QtCore.QCoreApplication.translate
-        CA.setWindowTitle(_translate("CA", "CA Experiment Creator"))
+        CV.setWindowTitle(_translate("CV", "CV Experiment Creator"))
         self.rest_time_label.setText(
             _translate(
-                "CA",
+                "CV",
                 '<html><head/><body><p align="center">Rest Time</p></body></html>',
             )
         )
         self.experiment_type_label.setText(
             _translate(
-                "CA",
+                "CV",
                 '<html><head/><body><p align="center">Experiment Type</p></body></html>',
-            )
-        )
-        self.experiment_step_number.setStatusTip(
-            _translate(
-                "CA",
-                "Number of steps between voltages over which the experiment will be conducted.",
             )
         )
         self.select_output_filepath_button.setStatusTip(
             _translate(
-                "CA",
+                "CV",
                 "Click this button to select the output filepath for your files.",
             )
         )
         self.select_output_filepath_button.setText(
-            _translate("CA", "Output Filepath")
+            _translate("CV", "Output Filepath")
         )
-        self.experiment_time_label.setText(
+        self.sweep_rate_label.setText(
             _translate(
-                "CA",
-                '<html><head/><body><p align="center">Experiment Time</p></body></html>',
+                "CV",
+                '<html><head/><body><p align="center">Sweep Rate</p></body></html>',
+            )
+        )
+        self.first_vertex_voltage_label.setText(
+            _translate(
+                "CV",
+                '<html><head/><body><p align="center">1st Vertex Voltage</p></body></html>',
             )
         )
         self.general_parameters_label.setText(
             _translate(
-                "CA",
+                "CV",
                 '<html><head/><body><p align="center">General Parameters</p></body></html>',
             )
         )
-        self.experiment_time.setStatusTip(
-            _translate("CA", "Time for which to conduct the experiment.")
-        )
         self.experiment_type_verify.setStatusTip(
-            _translate("CA", "Type of experiment for this file.")
+            _translate("CV", "Type of experiment for this file.")
         )
         self.experiment_type_verify.setText(
-            _translate("CA", "Chronoamperometry")
-        )
-        self.experiment_voltage.setStatusTip(
-            _translate("CA", "Voltage for experiment.")
+            _translate("CV", "Cyclic Voltammetry")
         )
         self.output_filename_label.setText(
             _translate(
-                "CA",
+                "CV",
                 '<html><head/><body><p align="center">Output Filename</p></body></html>',
             )
         )
         self.experiment_file_name.setStatusTip(
             _translate(
-                "CA",
+                "CV",
                 "Name that will be attached to. Data will be output to _data.csv and an experiment file will be saved to _config.yml.",
             )
         )
         self.step_number_label.setText(
             _translate(
-                "CA",
+                "CV",
                 '<html><head/><body><p align="center">Step Number</p></body></html>',
             )
         )
+        self.experiment_sweep_rate.setStatusTip(
+            _translate("CV", "Rate at which voltage is swept over.")
+        )
         self.experiment_parameters_label.setText(
             _translate(
-                "CA",
+                "CV",
                 '<html><head/><body><p align="center">Experiment Parameters</p></body></html>',
             )
         )
-        self.experiment_duration.setStatusTip(
-            _translate("CA", "Estimated length of experiment in H:M:S.")
-        )
         self.experiment_file_path.setStatusTip(
             _translate(
-                "CA",
+                "CV",
                 "Path at which data and experiment files will be exported to.",
             )
         )
         self.experiment_preview_label.setText(
             _translate(
-                "CA",
+                "CV",
                 '<html><head/><body><p align="center">Experiment Preview</p></body></html>',
             )
         )
-        self.experiment_duration_button.setText(
-            _translate(
-                "CA",
-                '<html><head/><body><p align="center">Experiment Duration</p></body></html>',
-            )
-        )
-        self.v_vs_label.setText(_translate("CA", "V vs."))
+        self.sweep_rate_units_label.setText(_translate("CV", "mV/s"))
+        self.v_vs_label_1.setText(_translate("CV", "V vs."))
+        self.v_vs_label_2.setText(_translate("CV", "V vs."))
         self.experiment_rest_time.setStatusTip(
-            _translate("CA", "Length of time to wait before start voltage.")
+            _translate("CV", "Length of time to wait before start voltage.")
         )
         self.advanced_parameters_button.setStatusTip(
             _translate(
-                "CA",
+                "CV",
                 "Click this button to edit hardset parameters used for all experiments.",
             )
         )
         self.advanced_parameters_button.setText(
-            _translate("CA", "Advanced Parameters")
+            _translate("CV", "Advanced Parameters")
         )
-        self.voltage_label.setText(
+        self.start_voltage_label.setText(
             _translate(
-                "CA",
-                '<html><head/><body><p align="center">Voltage</p></body></html>',
+                "CV",
+                '<html><head/><body><p align="center">Start Voltage</p></body></html>',
             )
         )
-        self.rest_time_units_label.setText(_translate("CA", "s"))
+        self.rest_time_units_label.setText(_translate("CV", "s"))
         self.save_experiment_file_button.setStatusTip(
             _translate(
-                "CA",
+                "CV",
                 "Click this button to save a config file with the created experiment.",
             )
         )
         self.save_experiment_file_button.setText(
-            _translate("CA", "Save Experiment File")
+            _translate("CV", "Save Experiment File")
         )
         self.generate_preview_button.setStatusTip(
             _translate(
-                "CA",
+                "CV",
                 "Click this button to generate a preview of the experiment to be run.",
             )
         )
         self.generate_preview_button.setText(
-            _translate("CA", "Generate Preview")
+            _translate("CV", "Generate Preview")
         )
-        self.rest_time_units_label_2.setText(_translate("CA", "s"))
-        self.voltage_ref.setCurrentText(_translate("CA", "Vref"))
-        self.voltage_ref.setItemText(0, _translate("CA", "Vref"))
-        self.voltage_ref.setItemText(1, _translate("CA", "Vocv"))
-        self.actionVref.setText(_translate("CA", "Vref"))
-        self.actionVocv.setText(_translate("CA", "Vocv"))
+        self.second_vertex_voltage_label.setText(
+            _translate(
+                "CV",
+                '<html><head/><body><p align="center">2nd Vertex Voltage</p></body></html>',
+            )
+        )
+        self.experiment_second_vertex_voltage.setStatusTip(
+            _translate(
+                "CV",
+                "Voltage at which to switch scan direction at on reverse scan.",
+            )
+        )
+        self.v_vs_label_3.setText(_translate("CV", "V vs."))
+        self.cycle_number_label.setText(
+            _translate(
+                "CV",
+                '<html><head/><body><p align="center">Cycle Number</p></body></html>',
+            )
+        )
+        self.experiment_cycle_number.setStatusTip(
+            _translate("CV", "Number of times to repeat cycle.")
+        )
+        self.experiment_step_number.setStatusTip(
+            _translate("CV", "Length of time to wait before start voltage.")
+        )
+        self.experiment_start_voltage.setStatusTip(
+            _translate("CV", "Length of time to wait before start voltage.")
+        )
+        self.experiment_first_vertex_voltage.setStatusTip(
+            _translate("CV", "Length of time to wait before start voltage.")
+        )
+        self.voltage_ref.setCurrentText(_translate("CV", "Vref"))
+        self.voltage_ref.setItemText(0, _translate("CV", "Vref"))
+        self.voltage_ref.setItemText(1, _translate("CV", "Vocv"))
+        self.voltage_ref_2.setCurrentText(_translate("CV", "Vref"))
+        self.voltage_ref_2.setItemText(0, _translate("CV", "Vref"))
+        self.voltage_ref_2.setItemText(1, _translate("CV", "Vocv"))
+        self.voltage_ref_3.setCurrentText(_translate("CV", "Vref"))
+        self.voltage_ref_3.setItemText(0, _translate("CV", "Vref"))
+        self.voltage_ref_3.setItemText(1, _translate("CV", "Vocv"))
+        self.experiment_duration_button.setText(
+            _translate(
+                "CV",
+                '<html><head/><body><p align="center">Experiment Duration</p></body></html>',
+            )
+        )
+        self.experiment_duration.setStatusTip(
+            _translate("CV", "Estimated length of experiment in H:M:S.")
+        )
 
 
 if __name__ == "__main__":
     import sys
 
     app = QtWidgets.QApplication(sys.argv)
-    CA = QtWidgets.QMainWindow()
-    ui = Ui_CA()
-    ui.setupUi(CA)
-    CA.show()
+    CV = QtWidgets.QMainWindow()
+    ui = Ui_CV()
+    ui.setupUi(CV)
+    CV.show()
     sys.exit(app.exec_())
