+++ src/pytentiostat/gui/GUI_LSV_exp_creator.py
@@ -8,7 +8,8 @@ from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.figure import Figure
 from PySide6.QtGui import QIcon
 from PySide6.QtWidgets import QGridLayout, QListWidgetItem
-from warning_GUI import warning
+
+from pytentiostat.gui.warning_GUI import warning
 
 
 # PartI: load output filepath
