+++ src/pytentiostat/gui/GUI_operator.py
@@ -2,10 +2,11 @@
 import datetime
 import time
 
+import numpy as np
+
 # Local library
 # GUI_function
-import GUI_config_reader as cr
-import numpy as np
+import pytentiostat.gui.GUI_config_reader as cr
 
 
 def start_exp(d9, normalized_start, data):
