+++ src/pytentiostat/gui/designer_files/Adv_Params.ui
