+++ src/pytentiostat/gui/designer_files/CA_gui.ui
