+++ src/pytentiostat/gui/designer_files/CV_gui.ui
