+++ src/pytentiostat/gui/designer_files/LSV.ui
