+++ src/pytentiostat/gui/designer_files/buffer.py
