+++ src/pytentiostat/gui/designer_files/exp_type.ui
