+++ src/pytentiostat/gui/designer_files/mainwindow.ui
