+++ src/pytentiostat/gui/designer_files/warning.ui
