+++ src/pytentiostat/gui/mainwindow_GUI.py
@@ -1,9 +1,10 @@
-from CA_GUI import Ui_CA
-from CV_GUI import Ui_CV
-from Exp_type_GUI import Ui_Experiment
-from GUI_load_config import Ui_Load
-from LSV_GUI import Ui_LSV
 from PySide6 import QtCore, QtGui, QtWidgets
+
+from pytentiostat.gui.CA_GUI import Ui_CA
+from pytentiostat.gui.CV_GUI import Ui_CV
+from pytentiostat.gui.Exp_type_GUI import Ui_Experiment
+from pytentiostat.gui.GUI_load_config import Ui_Load
+from pytentiostat.gui.LSV_GUI import Ui_LSV
 
 
 class Ui_MainWindow(object):
