+++ src/pytentiostat/pics/icon_ca.ico
