+++ src/pytentiostat/pics/icon_checked.ico
