+++ src/pytentiostat/pics/icon_connected.ico
