+++ src/pytentiostat/pics/icon_cv.ico
