+++ src/pytentiostat/pics/icon_disconnected.ico
