+++ src/pytentiostat/pics/icon_lsv.ico
