+++ src/pytentiostat/pics/icon_off.ico
