+++ src/pytentiostat/pics/icon_on.ico
