+++ src/pytentiostat/pics/icon_pause.ico
