+++ src/pytentiostat/pics/icon_pytentiostat.ico
