+++ src/pytentiostat/pics/icon_run.ico
