+++ src/pytentiostat/pics/icon_stop.ico
