+++ src/pytentiostat/pics/logo_pypot.png
