+++ src/pytentiostat/pics/logo_pytentiostat.png
